requirements.txt
ollama>=0.4.0
pystac-client>=0.8.0
rasterio>=1.3.0
pyproj>=3.6.0
Pillow>=10.0.0
numpy>=1.26.0
shapely>=2.0.0
streamlit>=1.38.0
